Last commit id:
commit ab3cca4ae16d05631f947def923645e6651e7779 (mirtos/master)
Author: liuyang34 <liuyang34@xiaomi.com>
Date:   Thu Sep 23 00:03:27 2021 -0700

    wapi: add 5g channel map in wapi

+++ include/netutils/netlib.h
@@ -1,5 +1,5 @@
 /****************************************************************************
- *  apps/include/netutils/netlib.h
+ * apps/include/netutils/netlib.h
  * Various non-standard APIs to support netutils.  All non-standard and
  * intended only for internal use.
  *
@@ -225,7 +225,7 @@ struct url_s
 #endif
   FAR char *host;
   int       hostlen;
-  int       port;
+  uint16_t  port;
   FAR char *path;
   int       pathlen;
 #if 0 /* not yet */
